requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==1.10.13
python-dotenv==1.0.1
databricks-sql-connector==3.1.0
pandas==2.1.4
joblib==1.3.2
scikit-learn==1.4.0
pytest==8.0.0
requests==2.32.3
folium==0.17.0
streamlit-folium==0.22.0